This inlcude CRUD in fastapi

+++ frontend/delete.js
@@ -2,16 +2,34 @@ document.addEventListener("DOMContentLoaded", function(){
 
     const deletebyidbutton = document.querySelector(".deletebyid");
     deletebyidbutton.addEventListener("click", async function(){
-        const id = document.querySelector("#idInput").value;
-        const response = await fetch(`http://127.0.0.1:8000/Passengers/${id}`,{
+        const idInput = document.querySelector("#idInput");
+        const id = idInput.value;
+        const displaydiv = document.querySelector("#deletepassengerbyid");
+        if (!idInput.value.trim()){
+            idInput.setCustomValidity("Please enter an Id first");
+            idInput.reportValidity();
+            return
+        }
+        idInput.setCustomValidity("");
+        try{
+            const response = await fetch(`http://127.0.0.1:8000/Passengers/${id}`,{
             method:"DELETE"
         });
+        
+        if(!response.ok){
+            const error = await response.json();
+            confirm(error.detail);
+            return;
+        }
         const data = await response.json();
-        const displaydiv = document.querySelector("#deletepassengerbyid");
-
-        displaydiv.innerHTML = `
-        <p>Passenger Deleted Successfully</p>
-        `;
+        displaydiv.innerHTML = `<p>${data.message}</p>`;
+        setTimeout(() =>{
+                displaydiv.innerHTML = "";
+            }, 3000);
+        } catch(error){
+            displaydiv.innerHTML = `
+            <p>${error.message}</p>
+            `;
+        }
     });
-
 });
